Add formatter to language server
