+++ common/changes/@cadl-lang/compiler/feature-lsp-formatter_2022-09-07-12-57.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@cadl-lang/compiler",
+      "comment": "Language server provide document formatting using Cadl formatter",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@cadl-lang/compiler"
+}
